+++ src/app-sdk/winrt/MidiEndpointConnection_SendInfrastructure.cpp
@@ -59,7 +59,7 @@ namespace winrt::Microsoft::Windows::Devices::Midi2::implementation
     midi2::MidiSendMessageResults MidiEndpointConnection::SendMessageRaw(
         winrt::com_ptr<IMidiBidirectional> endpoint,
         void* data,
-        uint8_t sizeInBytes,
+        uint32_t sizeInBytes,
         internal::MidiTimestamp timestamp)
     {
 #ifdef _DEBUG
